Board class basic outline
Added Constructors, Setters, Getters, and toString method to the board class, modified the original outputBoard calls from the TicTacToeComputer class. Packaged in a "tictactoe" package.
